finalizando ui registros

+++ consultoriomedicobusiness/src/main/java/edu/itla/consultoriomedico/business/repository/PacienteRepository.java
@@ -2,5 +2,5 @@ package edu.itla.consultoriomedico.business.repository;
 
 import edu.itla.consultoriomedico.business.entity.Paciente;
 
-public interface PacienteRepository extends GenericRepository<Paciente, Integer>{
+public interface PacienteRepository extends GenericRepository<Paciente, Long>{
 }
